Notice that the flash_config structure is incorrectly defined within the picow_ntp_client.c file. Change the code so that we only have one definition. Move things into .h files now that we have the Pico-Green-Clock.h file to use.
